Этап 2. основной сервис Основной сервис без докера

+++ ewm-service/src/main/java/ru/practicum/service/ParticipationRequestService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ParticipationRequestService {
     List<ParticipationRequestDto> getUserRequests(Long userId);
 
-    ParticipationRequestDto createRequest(Long userId, Integer eventId);
+    ParticipationRequestDto createRequest(Long userId, Long eventId);
 
     ParticipationRequestDto cancelRequest(Long userId, Long requestId);
 }
